+++ mattermost-plugin/server/plugin.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/url"
 	"path"
 	"strings"
 	"sync"
@@ -22,7 +20,6 @@ import (
 
 	mmModel "github.com/mattermost/mattermost-server/v6/model"
 	"github.com/mattermost/mattermost-server/v6/plugin"
-	"github.com/mattermost/mattermost-server/v6/shared/markdown"
 	"github.com/mattermost/mattermost-server/v6/shared/mlog"
 )
 
@@ -31,15 +28,6 @@ const (
 	pluginName            = "focalboard"
 	sharedBoardsName      = "enablepublicsharedboards"
 )
-
-type BoardsEmbed struct {
-	OriginalPath string `json:"originalPath"`
-	WorkspaceID  string `json:"workspaceID"`
-	ViewID       string `json:"viewID"`
-	BoardID      string `json:"boardID"`
-	CardID       string `json:"cardID"`
-	ReadToken    string `json:"readToken,omitempty"`
-}
 
 // Plugin implements the interface expected by the Mattermost server to communicate between the server and plugin processes.
 type Plugin struct {
@@ -257,128 +245,17 @@ func defaultLoggingConfig() string {
 }
 
 func (p *Plugin) MessageWillBePosted(_ *plugin.Context, post *mmModel.Post) (*mmModel.Post, string) {
-	return postWithBoardsEmbed(post, p.API.GetConfig().FeatureFlags.BoardsUnfurl), ""
+	post, err := postWithBoardsEmbed(post, p.API.GetConfig().FeatureFlags.BoardsUnfurl)
+	if err != nil {
+		p.server.Logger().Warn("Error embedding link in post", mlog.Err(err))
+	}
+	return post, ""
 }
 
 func (p *Plugin) MessageWillBeUpdated(_ *plugin.Context, newPost, _ *mmModel.Post) (*mmModel.Post, string) {
-	return postWithBoardsEmbed(newPost, p.API.GetConfig().FeatureFlags.BoardsUnfurl), ""
-}
-
-func postWithBoardsEmbed(post *mmModel.Post, showBoardsUnfurl bool) *mmModel.Post {
-	if _, ok := post.GetProps()["boards"]; ok {
-		post.AddProp("boards", nil)
-	}
-
-	if !showBoardsUnfurl {
-		return post
-	}
-
-	firstLink := getFirstLink(post.Message)
-
-	if firstLink == "" {
-		return post
-	}
-
-	u, err := url.Parse(firstLink)
-
+	post, err := postWithBoardsEmbed(newPost, p.API.GetConfig().FeatureFlags.BoardsUnfurl)
 	if err != nil {
-		return post
+		p.server.Logger().Warn("Error embedding link in post", mlog.Err(err))
 	}
-
-	// Trim away the first / because otherwise after we split the string, the first element in the array is a empty element
-	urlPath := u.Path
-	if strings.HasPrefix(urlPath, "/") {
-		urlPath = u.Path[1:]
-	}
-	pathSplit := strings.Split(strings.ToLower(urlPath), "/")
-	queryParams := u.Query()
-
-	if len(pathSplit) == 0 {
-		return post
-	}
-
-	workspaceID, boardID, viewID, cardID := returnBoardsParams(pathSplit)
-
-	if workspaceID != "" && boardID != "" && viewID != "" && cardID != "" {
-		b, _ := json.Marshal(BoardsEmbed{
-			WorkspaceID:  workspaceID,
-			BoardID:      boardID,
-			ViewID:       viewID,
-			CardID:       cardID,
-			ReadToken:    queryParams.Get("r"),
-			OriginalPath: u.RequestURI(),
-		})
-
-		BoardsPostEmbed := &mmModel.PostEmbed{
-			Type: mmModel.PostEmbedBoards,
-			Data: string(b),
-		}
-
-		if post.Metadata == nil {
-			post.Metadata = &mmModel.PostMetadata{}
-		}
-
-		post.Metadata.Embeds = []*mmModel.PostEmbed{BoardsPostEmbed}
-		post.AddProp("boards", string(b))
-	}
-
-	return post
-}
-
-func getFirstLink(str string) string {
-	firstLink := ""
-
-	markdown.Inspect(str, func(blockOrInline interface{}) bool {
-		if _, ok := blockOrInline.(*markdown.Autolink); ok {
-			if link := blockOrInline.(*markdown.Autolink).Destination(); firstLink == "" {
-				firstLink = link
-			}
-		}
-		return true
-	})
-
-	return firstLink
-}
-
-func returnBoardsParams(pathArray []string) (workspaceID, boardID, viewID, cardID string) {
-	// The reason we are doing this search for the first instance of boards or plugins is to take into account URL subpaths
-	index := -1
-	for i := 0; i < len(pathArray); i++ {
-		if pathArray[i] == "boards" || pathArray[i] == "plugins" {
-			index = i
-			break
-		}
-	}
-
-	if index == -1 {
-		return workspaceID, boardID, viewID, cardID
-	}
-
-	// If at index, the parameter in the path is boards,
-	// then we've copied this directly as logged in user of that board
-
-	// If at index, the parameter in the path is plugins,
-	// then we've copied this from a shared board
-
-	// For card links copied on a non-shared board, the path looks like {...Mattermost Url}.../boards/workspace/workspaceID/boardID/viewID/cardID
-
-	// For card links copied on a shared board, the path looks like
-	// {...Mattermost Url}.../plugins/focalboard/workspace/workspaceID/shared/boardID/viewID/cardID?r=read_token
-
-	// This is a non-shared board card link
-	if len(pathArray)-index == 6 && pathArray[index] == "boards" && pathArray[index+1] == "workspace" {
-		workspaceID = pathArray[index+2]
-		boardID = pathArray[index+3]
-		viewID = pathArray[index+4]
-		cardID = pathArray[index+5]
-	} else if len(pathArray)-index == 8 && pathArray[index] == "plugins" &&
-		pathArray[index+1] == "focalboard" &&
-		pathArray[index+2] == "workspace" &&
-		pathArray[index+4] == "shared" { // This is a shared board card link
-		workspaceID = pathArray[index+3]
-		boardID = pathArray[index+5]
-		viewID = pathArray[index+6]
-		cardID = pathArray[index+7]
-	}
-	return workspaceID, boardID, viewID, cardID
+	return post, ""
 }
